Update airbase to 253
(cherry picked from commit ef6b6a92bee5942f7c144ece5c4ac5d0970e4056)
(cherry picked from commit 50633719ce828f8d9c4fe3d3ab5d932a37383628)

+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/test/java/io/trino/plugin/warp/juffer/PredicatesCacheServiceTest.java
@@ -98,7 +98,7 @@ public class PredicatesCacheServiceTest
     {
         PredicateData predicateData = buildMockedPredicateData(MemorySegment.NULL, predicateBufferPoolType, 1);
         ArgumentCaptor<PredicateCacheData> argument = ArgumentCaptor.forClass(PredicateCacheData.class);
-        Domain domain = Domain.singleValue(IntegerType.INTEGER, (long) 1);
+        Domain domain = Domain.singleValue(IntegerType.INTEGER, 1L);
         PredicateCacheData actualBufferHandle = predicatesCacheService.getOrCreatePredicateBufferId(predicateData, domain).orElseThrow();
         assertThat(actualBufferHandle.isUsed()).isTrue();
         predicatesCacheService.markFinished(List.of(actualBufferHandle));
@@ -113,8 +113,8 @@ public class PredicatesCacheServiceTest
         PredicateData predicateData = buildMockedPredicateData(MemorySegment.NULL, predicateBufferPoolType, 1);
         PredicateData predicateData2 = buildMockedPredicateData(MemorySegment.NULL, predicateBufferPoolType, 2);
         ArgumentCaptor<PredicateCacheData> argument = ArgumentCaptor.forClass(PredicateCacheData.class);
-        Domain domain = Domain.singleValue(IntegerType.INTEGER, (long) 1);
-        Domain domain2 = Domain.singleValue(IntegerType.INTEGER, (long) 2);
+        Domain domain = Domain.singleValue(IntegerType.INTEGER, 1L);
+        Domain domain2 = Domain.singleValue(IntegerType.INTEGER, 2L);
         PredicateCacheData actualBufferHandle;
         actualBufferHandle = predicatesCacheService.getOrCreatePredicateBufferId(predicateData, domain).orElseThrow();
         assertThat(actualBufferHandle.isUsed()).isTrue();
@@ -139,7 +139,7 @@ public class PredicatesCacheServiceTest
     public void testNoFreeEntriesPoolMedium()
     {
         PredicateData predicateData = buildMockedPredicateData(MemorySegment.NULL, PredicateBufferPoolType.MEDIUM, 1);
-        Domain domain = Domain.singleValue(IntegerType.INTEGER, (long) 1);
+        Domain domain = Domain.singleValue(IntegerType.INTEGER, 1L);
         when(bufferAllocator.getPoolSize(PredicateBufferPoolType.MEDIUM)).thenReturn(0);
         Optional<PredicateCacheData> actualBufferHandle = predicatesCacheService.getOrCreatePredicateBufferId(predicateData, domain);
         assertThat(actualBufferHandle).isEmpty();
